+++ cmd/update.go
@@ -55,7 +55,11 @@ func UpdateServer(root *cmd.Root, c *cmd.Sub) {
 
 	Log.Infoln("Downloading new server jar...")
 	if err := prov.Download(outFile); err != nil {
-		Log.Fatalln("Error downloading file:", err)
+		if err == provider.ErrAlreadyUpToDate {
+			Log.Goodln("Server jar is already up to date")
+		} else {
+			Log.Fatalln("Error downloading file:", err)
+		}
 	} else {
 		Log.Goodln("Server jar updated!")
 	}