+++ provider/paper.go
@@ -2,8 +2,10 @@ package provider
 
 import (
 	"encoding/json"
+	"errors"
 	"fmt"
 	"net/http"
+	"os"
 
 	"github.com/stretchr/stew/slice"
 )
@@ -15,28 +17,59 @@ const (
 	paperDownloadEndpoint = "https://papermc.io/api/v2/projects/paper/versions/%s/builds/%d/downloads/%s"
 )
 
-// getLatestBuild queries the Paper API to get the latest build number for the
-// version of Minecraft we were given.
-func (p Paper) getLatestBuild() (int, error) {
-	url := fmt.Sprintf(paperVersionsEndpoint, p.Version)
-	resp, err := http.Get(url)
+// ErrAErrAlreadyUpToDate is an error returned when the server is already
+// at the latest build for the given version.
+var ErrAlreadyUpToDate = errors.New("paper version is already at the latest build")
+
+// Paper is an update provider that downloads a new Paper server version.
+type Paper struct {
+	Version string
+}
+
+// Download gets the latest build of Paper from their website
+// for the given Minecraft version.
+func (p Paper) Download(filepath string) error {
+	// See if we actially have a valid version
+	valid, err := p.validateVersion()
 	if err != nil {
-		return 0, err
+		return err
 	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != 200 {
-		return 0, fmt.Errorf("failed to get builds for version '%s': %d", p.Version, resp.StatusCode)
+	if !valid {
+		return fmt.Errorf("server version not found: %s", p.Version)
 	}
 
-	dec := json.NewDecoder(resp.Body)
-	builds := &PaperBuilds{}
-	err = dec.Decode(builds)
+	// Get the latest build number
+	b, err := getLatestBuild(p.Version)
 	if err != nil {
-		return 0, err
+		return err
 	}
 
-	return builds.Builds[len(builds.Builds)-1], nil
+	// Check if we have the version we're currently running saved
+	saved, err := Load(".paper_build.json")
+	if err != nil {
+		return fmt.Errorf("unable to read old version: %s", err.Error())
+	}
+
+	// Check if the current version and build matches the latest
+	if saved.Version == b.Version {
+		if saved.Build == b.Build {
+			return ErrAlreadyUpToDate
+		}
+	}
+
+	// Download the actual jar file
+	url := fmt.Sprintf(paperDownloadEndpoint, p.Version, b.Build, b.Download.Application.Name)
+	if err = DownloadFile(url, filepath); err != nil {
+		return err
+	}
+
+	// Save the new version and build to disk
+	if err = b.Save(".paper_build.json"); err != nil {
+		return fmt.Errorf("unable to save version file: %s", err.Error())
+	}
+
+	// Verify the downloaded file
+	return Verify(filepath, b.Download.Application.Hash)
 }
 
 // validateVersion queries the Paper API to see if we have a valid version string.
@@ -61,45 +94,113 @@ func (p Paper) validateVersion() (bool, error) {
 	return slice.Contains(versions.Versions, p.Version), nil
 }
 
-// Download gets the latest build of Paper from their website
-// for the given Minecraft version.
-func (p Paper) Download(filepath string) error {
-	// See if we actially have a valid version
-	valid, err := p.validateVersion()
+// PaperVersions is the representation of all Paper versions returned by the API.
+type PaperVersions struct {
+	VersionGroups []string `json:"version_groups"`
+	Versions      []string `json:"versions"`
+}
+
+// PaperBuilds is the representation of the Paper API response for a version.
+type PaperBuilds struct {
+	Builds []int `json:"builds"`
+}
+
+// PaperBuild holds the API response data for a particular Paper build.
+type PaperBuild struct {
+	Build    int           `json:"build"`
+	Download PaperDownload `json:"downloads"`
+	Version  string        `json:"version"`
+}
+
+// PaperDownload contains information about a build's file.
+type PaperDownload struct {
+	Application PaperApplication `json:"application"`
+}
+
+// PaperApplication holds the name and hash of a file for a Paper build.
+type PaperApplication struct {
+	Name string `json:"name"`
+	Hash string `json:"sha256"`
+}
+
+// Load reads saved version information from a file.
+// If the file does not exist, this function returns
+// an empty struct and no error.
+func Load(path string) (*PaperBuild, error) {
+	if _, err := os.Stat(path); err != nil {
+		if errors.Is(err, os.ErrNotExist) {
+			return &PaperBuild{}, nil
+		}
+		return nil, err
+	}
+
+	file, err := os.Open(path)
+	if err != nil {
+		return nil, err
+	}
+	defer file.Close()
+
+	b := PaperBuild{}
+	dec := json.NewDecoder(file)
+	if err = dec.Decode(&b); err != nil {
+		return nil, err
+	}
+
+	return &b, nil
+}
+
+// Save write a Paper build to a file on disk.
+func (p PaperBuild) Save(path string) error {
+	file, err := os.Create(path)
 	if err != nil {
 		return err
 	}
-	if !valid {
-		return fmt.Errorf("server version not found: %s", p.Version)
-	}
+	defer file.Close()
 
-	// Get the latest build number
-	build, err := p.getLatestBuild()
-	if err != nil {
-		return err
-	}
+	return json.NewEncoder(file).Encode(p)
+}
 
-	url := fmt.Sprintf(paperBuildEndpoint, p.Version, build)
+// getLatestBuild queries the Paper API to get the latest build for the
+// version of Minecraft we were given.
+func getLatestBuild(version string) (*PaperBuild, error) {
+	// Get the list of builds for the given version
+	url := fmt.Sprintf(paperVersionsEndpoint, version)
 	resp, err := http.Get(url)
 	if err != nil {
-		return err
+		return nil, err
 	}
 	defer resp.Body.Close()
 
 	if resp.StatusCode != 200 {
-		return fmt.Errorf("failed to get build info: %d", resp.StatusCode)
+		return nil, fmt.Errorf("failed to get builds for version '%s': %d", version, resp.StatusCode)
 	}
 
 	dec := json.NewDecoder(resp.Body)
+	builds := &PaperBuilds{}
+	err = dec.Decode(builds)
+	if err != nil {
+		return nil, err
+	}
+
+	build := builds.Builds[len(builds.Builds)-1]
+
+	// Get the latest build info for this version
+	url = fmt.Sprintf(paperBuildEndpoint, version, build)
+	buildResp, err := http.Get(url)
+	if err != nil {
+		return nil, err
+	}
+	defer buildResp.Body.Close()
+
+	if resp.StatusCode != 200 {
+		return nil, fmt.Errorf("failed to get build info: %d", buildResp.StatusCode)
+	}
+
+	dec = json.NewDecoder(buildResp.Body)
 	var b PaperBuild
 	if err = dec.Decode(&b); err != nil {
-		return err
+		return nil, err
 	}
 
-	url = fmt.Sprintf(paperDownloadEndpoint, p.Version, build, b.Download.Application.Name)
-	if err = DownloadFile(url, filepath); err != nil {
-		return err
-	}
-
-	return Verify(filepath, b.Download.Application.Hash)
+	return &b, nil
 }